changing the type of all id

+++ src/project/Layer/layer.cpp
@@ -1,6 +1,6 @@
 #include "layer.hpp"
 
-Layer::Layer(uint64_t id,std::string name, sf::Vector2u size) : id_{id}, name_{name}, size_{size} {}
+Layer::Layer(uint id,std::string name, sf::Vector2u size) : id_{id}, name_{name}, size_{size} {}
 
 
 std::string Layer::getName() const { return name_; }
