+++ src/project/Layer/spritelayer.cpp
@@ -1,6 +1,6 @@
 #include "spritelayer.hpp"
 
-SpriteLayer::SpriteLayer(uint64_t id, std::string name, sf::Vector2u size)
+SpriteLayer::SpriteLayer(uint id, std::string name, sf::Vector2u size)
     : Layer(id,name, size) {
   type_ = SPRITELAYER;
 }
